# tests/unit/openapi_snapshot.json
{
  "components": {
    "schemas": {
...
      "AdminEmployerListResponse": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/AdminEmployerRead"
            },
            "title": "Items",
            "type": "array"
          },
          "next_cursor": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Next Cursor"
          }
        },
        "required": [
          "items"
        ],
        "title": "AdminEmployerListResponse",
        "type": "object"
      },
      "AdminEmployerRead": {
        "description": "Admin view of an employer's verification state. `reviewed_at`/`reason`\nare derived \u2014 there is no separate review table; `audit_logs` is the history.",
        "properties": {
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "gst": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Gst"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "reason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reason"
          },
          "reviewed_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reviewed At"
          },
          "status": {
            "enum": [
              "pending",
              "verified",
              "rejected"
            ],
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "id",
          "name",
          "gst",
          "status",
          "created_at",
          "reviewed_at",
          "reason"
        ],
        "title": "AdminEmployerRead",
        "type": "object"
      },
...
      "EmployerCreate": {
        "additionalProperties": false,
        "properties": {
          "gst": {
            "anyOf": [
              {
                "maxLength": 15,
                "minLength": 15,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Gst"
          },
          "name": {
            "maxLength": 200,
            "minLength": 2,
            "title": "Name",
            "type": "string"
          }
        },
        "required": [
          "name"
        ],
        "title": "EmployerCreate",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "RejectEmployerRequest": {
        "additionalProperties": false,
        "properties": {
          "reason": {
            "maxLength": 255,
            "minLength": 1,
            "title": "Reason",
            "type": "string"
          }
        },
        "required": [
          "reason"
        ],
        "title": "RejectEmployerRequest",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
      "jobify_api__routes__employers__core__EmployerRead": {
        "properties": {
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "gst": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Gst"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "verified_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Verified At"
          }
        },
        "required": [
          "id",
          "name",
          "gst",
          "verified_at",
          "created_at"
        ],
        "title": "EmployerRead",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "Jobify API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/v1/admin/employers": {
      "get": {
        "operationId": "list_employers_for_verification_v1_admin_employers_get",
        "parameters": [
          {
            "in": "query",
            "name": "status",
            "required": false,
            "schema": {
              "default": "pending",
              "enum": [
                "pending",
                "verified",
                "rejected"
              ],
              "title": "Status",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "cursor",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Cursor"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 20,
              "maximum": 100,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminEmployerListResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Employers For Verification",
        "tags": [
          "admin"
        ]
      }
    },
...
    "/v1/admin/employers/{employer_id}/reject": {
      "post": {
        "operationId": "reject_employer_v1_admin_employers__employer_id__reject_post",
        "parameters": [
          {
            "in": "path",
            "name": "employer_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Employer Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RejectEmployerRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminEmployerRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Reject Employer",
        "tags": [
          "admin"
        ]
      }
    },
    "/v1/admin/employers/{employer_id}/verify": {
      "post": {
        "operationId": "verify_employer_v1_admin_employers__employer_id__verify_post",
        "parameters": [
          {
            "in": "path",
            "name": "employer_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Employer Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminEmployerRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Verify Employer",
        "tags": [
          "admin"
        ]
      }
    },
...
    "/v1/employers": {
      "post": {
        "operationId": "create_employer_v1_employers_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EmployerCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/jobify_api__routes__employers__core__EmployerRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Employer",
        "tags": [
          "employers"
        ]
      }
    },
...
  }
}